+++ examples/web/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @workglow/web
 
+## 0.0.54
+
+### Patch Changes
+
+- Update output shcema on input changes for FetchTask
+- Updated dependencies
+  - @workglow/ai-provider@0.0.54
+  - @workglow/task-graph@0.0.54
+  - @workglow/job-queue@0.0.54
+  - @workglow/storage@0.0.54
+  - @workglow/sqlite@0.0.54
+  - @workglow/debug@0.0.54
+  - @workglow/tasks@0.0.54
+  - @workglow/test@0.0.54
+  - @workglow/util@0.0.54
+  - @workglow/ai@0.0.54
+
+## 0.0.53
+
+### Patch Changes
+
+- Update FetchTask to use dynamic output schema
+- Updated dependencies
+  - @workglow/tasks@0.0.53
+  - @workglow/test@0.0.53
+  - @workglow/ai@0.0.53
+  - @workglow/ai-provider@0.0.53
+  - @workglow/debug@0.0.53
+  - @workglow/job-queue@0.0.53
+  - @workglow/sqlite@0.0.53
+  - @workglow/storage@0.0.53
+  - @workglow/task-graph@0.0.53
+  - @workglow/util@0.0.53
+
 ## 0.0.52
 
 ### Patch Changes