+++ packages/sqlite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @workglow/sqlite
 
+## 0.0.54
+
+### Patch Changes
+
+- Update output shcema on input changes for FetchTask
+
+## 0.0.53
+
+### Patch Changes
+
+- Update FetchTask to use dynamic output schema
+
 ## 0.0.52
 
 ### Patch Changes