+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @workglow/storage
 
+## 0.0.54
+
+### Patch Changes
+
+- Update output shcema on input changes for FetchTask
+- Updated dependencies
+  - @workglow/sqlite@0.0.54
+  - @workglow/util@0.0.54
+
+## 0.0.53
+
+### Patch Changes
+
+- Update FetchTask to use dynamic output schema
+- Updated dependencies
+  - @workglow/sqlite@0.0.53
+  - @workglow/util@0.0.53
+
 ## 0.0.52
 
 ### Patch Changes